requirements.txt
# Dependencias principales del proyecto AbsentismoEspana
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
jinja2>=3.0.0
openpyxl>=3.0.0
requests>=2.25.0
psycopg2-binary>=2.9.0
python-docx>=0.8.11
chardet>=4.0.0
